+++ packages/forms/src/FormSkeleton.js
@@ -1,30 +1,29 @@
 import React from 'react';
+
 import PropTypes from 'prop-types';
-import Skeleton from '@talend/react-components/lib/Skeleton';
+
+import { SkeletonInput, SkeletonButton } from '@talend/design-system';
+
 import theme from './FormSkeleton.scss';
 
 export default function FormSkeleton({ displayMode, actions }) {
 	// null/undefined actions prop will display default buttons
 	const hasButtons = displayMode !== 'text' && actions?.length !== 0;
 	return (
-		<div className={`${theme.container} tc-skeleton-heartbeat`} aria-busy>
-			<div className={theme['form-content']}>
+		<div className={theme.container} aria-busy data-testid="form.skeleton">
+			<div className={theme['form-content']} data-testid="form.skeleton.fields">
 				<div className={theme['form-content-wrapper']}>
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} size={Skeleton.SIZES.large} />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} className="skeleton-fit-content" />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} size={Skeleton.SIZES.large} />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} className="skeleton-fit-content" />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} size={Skeleton.SIZES.large} />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} className="skeleton-fit-content" />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} size={Skeleton.SIZES.large} />
-					<Skeleton heartbeat={false} type={Skeleton.TYPES.text} className="skeleton-fit-content" />
+					<SkeletonInput />
+					<SkeletonInput />
+					<SkeletonInput />
+					<SkeletonInput />
 				</div>
 			</div>
 			{hasButtons && (
-				<div className={theme.submit}>
+				<div className={theme.submit} data-testid="form.skeleton.buttons">
 					<div className={theme['submit-wrapper']}>
-						<Skeleton heartbeat={false} type={Skeleton.TYPES.button} />
-						<Skeleton heartbeat={false} type={Skeleton.TYPES.button} />
+						<SkeletonButton />
+						<SkeletonButton />
 					</div>
 				</div>
 			)}